requirements.txt
# local package
# -e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
pyspark
notebook
numpy
pandas
fastparquet
pyarrow
findspark
pandarallel
apache-airflow[yandexcloud]
apache-airflow-providers-yandex
venv-pack
